lambda/python/import_truck_stats/requirements.txt
google-cloud-storage==1.13.2
google-cloud-datastore==1.7.3

